+++ threadx_test/evaluation/evaluation_files/byte_pool_create.c
@@ -4,6 +4,8 @@
 
 extern TX_THREAD *    _tx_thread_current_ptr;
 
+UCHAR                   pool_start[9120];
+
 int main()
 {
     klee_make_symbolic_controllable(&_tx_thread_current_ptr, sizeof(_tx_thread_current_ptr), "_tx_thread_current_ptr", false);
@@ -12,8 +14,6 @@ int main()
     klee_make_symbolic_controllable(&pool_ptr, sizeof(pool_ptr), "pool_ptr", true);
     CHAR * name_ptr;
     klee_make_symbolic_controllable(&name_ptr, sizeof(name_ptr), "name_ptr", true);
-    VOID * pool_start;
-    klee_make_symbolic_controllable(&pool_start, sizeof(pool_start), "pool_start", true);
     ULONG pool_size;
     klee_make_symbolic_controllable(&pool_size, sizeof(pool_size), "pool_size", true);
     UINT pool_control_block_size;